ai/requirements.txt
annotated-types
anyio
beautifulsoup4
cachetools
certifi
charset-normalizer
click
fastapi
filelock
fsspec
google-ai-generativelanguage
google-api-core
google-auth
google-generativeai
googleapis-common-protos
google-cloud-texttospeech
google-cloud-speech
grpcio
grpcio-status
h11
huggingface-hub
idna
Jinja2
Markdown
MarkupSafe
packaging
proto-plus
protobuf
pyasn1
pyasn1-modules
pydantic
pydantic_core
PyYAML
requests
rsa
sniffio
soupsieve
starlette
tqdm
typing_extensions
urllib3
uvicorn
python-multipart
markdown2
openai==0.28
sqlalchemy
bcrypt
httpx
PyMySQL
typing_extensions
cryptography
torch
torchvision
mediapipe
librosa
boto3
pyaudio
python-jose
itsdangerous
aiomysql
asyncmy
moviepy
pydub